fix(stable): tests not passing
Last Rust update complains about the resolver not being explicit: `some crates are on edition 2021 which defaults to `resolver = "2"`, but virtual workspaces default to `resolver = "1"`

On the other hand, when tried to apply this commit, got some other errors from clippy which invalidated the commit.

I changed the usage of an `Arc` which was wrapping a `Repository` which was not `Send` nor `Sync`.

Haven't looked much into the details but it doesn't seem that this needs to be `Send` or `Sync` so I just opted to use a plain `Rc`.