Parse improperly formatted YAML for the ruleset config.
In the first approach to this, we decided to only read the ruleset
config as a list of strings or as a map ("proper YAML"). However,
we have observed that many people, when they convert their old
configurations, build it as lists of maps.

In fact, when people write up examples of configurations from scratch,
they often do that, and they are not even converting old configurations.

This suggests that we really should read these configurations and parse
them as the user expects, even if they are not syntactically proper.

This commit changes the parser so it will be more forgiving and accept
lists of maps as valid configuration.

Before, only these were accepted:

rulesets:
  - rs_one
  - rs_two

rulesets:
  rs_one:
  rs_two:
    only:
      - foo

Now, these will be accepted as well:

rulesets:
  - rs_one:
    only:
      - foo
  - rs_two:
      only:
        -foo

These will still not be accepted:

rulesets:
  - rs_one
    only:
      - foo
  - rs_two
      only:
        - foo