feat(kernel): add support for ConfigFile serialization (#263)
## What problem are you trying to solve?

I was trying to serialize a ConfigFile instance and expected to get a valid YAML file. Unfortunately, the serialization for the rulesets was not implemented resulting in an invalid YAML according to the schema of our configuration file.

## What is your solution?

This PR adds support for proper serialization, ignoring both `Option::None` and some default values. This helps to avoid having values showing `null`, `{}` or `[]`

## Alternatives considered

## What the reviewer should know

Added a few tests for the new functionality and a vscode debug configuration to help debugging the tests, if needed.

Also, I added a new crate called [indexmap](https://crates.io/crates/indexmap), which was needed to keep the insertion order of the keys in the maps. This is critical for serialization purposes as the `rulesets` array has a weird convention where the first key must be the ruleset name with a null value.