feat(config): preserve comments in the static analysis configuration file - IDE-2599 (#472)
The last commit fixes some clippy warnings that were triggering when committing or pushing (via cargo-husky). I had to touch some files outside the ide folder. For some reason, we're using -D warnings locally but not in the CI. I didn't want to modify any of this but I guess it would be nice to reconcile this difference at some point.

IDE-2599