feat(server): add option to save logs to a file (#541)
## What problem are you trying to solve?

We have been observing several issues when using the server in the IDE:
- intelliJ reports having multiple instances of the server that won't be automatically shutdown
- a user in VS Code reported a case where multiple processes where present at the same time after leaving his computer idle all the night (unable to reproduce on our side)

## What is your solution?

In order to better understand what's going on it would be nice to have the generated logs of the server saved into a file.

This PR adds a new option in the CLI to roll all the logs to a file. The user can decide the cadence: minutely, hourly or daily. These logs will be using JSON format for easy parsing.

These logs are saved into the `/tmp` folder in UNIX and will default to the `TEMP` env var in windows, which is usually pointing to `C:\Users\<user_name>\AppData\Local\Temp`. If not present, for whatever reason, then it will fallback to `c:\Temp`.

The idea is to leverage this option to enable a setting in the IDEs that will allow users and the dev team to enable this feature in the binary and then easily debug the static analysis server processes. 

## Alternatives considered

## What the reviewer should know

Tested it on Windows 11 and it works, thought the compilation process didn't work out of the box due to some UNIX commands being used in the kernel build.rs file and a wrong dependency version. I managed to fix all that. I'll prepare another PR for that.


IDE-3768