fix(server): better tracing management (#543)
## What problem are you trying to solve?

There were different threads using process::exit. This was making it difficult to correctly flush the tracing messages.

On top of that, the `WorkerGuard` was shared but this was introducing other issues, as multiple references would avoid flushing the traces.

In conclusion, the mixture of these two facts was making the tracing messages not being correctly flushed

## What is your solution?

Centralized all the panics into the main thread so we can guarantee that the tracing messages are properly flushed by dropping the `WorkerGuard` if needed.

## Alternatives considered

## What the reviewer should know

There's only one panic left in the keep alive thread. This is a last resource mechanism, and it's only being used after having tried 2 different alternatives before.

While at the task I added errors 100, 102 and 103 so we can have more useful information if we find those codes in the IDE and handle those cases accordingly.

IDE-3768