Fix docker run command
The command should have the volume mount before the image name. Arguments after the image name are sent to the application within the container.
Also added '-it' and '--rm' so the container cleans itself and has a interactive TTY to print to. It seems like a more sensible copy-past default to me