feat(ide): avoid GET url length issues
Added a v2 version for some GET endpoints that were expecting a base64 representation of a file content as URL path.

We also added a new version for the POST rulesets endpoint just because the naming was more convenient.

The idea is to remove the v1 endpoints in 6 months (or re-evaluate).

IDE-4957