feat(ide): avoid GET url length issues (#756)
## What problem are you trying to solve?

See https://github.com/DataDog/datadog-static-analyzer/pull/754 for more details.

Essentially, the problem we may encounter is URL length being a problem for some of the GET endpoints.

## What is your solution?

Added a v2 version for some GET endpoints that were expecting a base64 representation of a file content as URL path.

We also added a new version for the POST rulesets endpoint just because the naming was more convenient.

## Alternatives considered

## What the reviewer should know

This is a continuation of https://github.com/DataDog/datadog-static-analyzer/pull/754

The idea is to remove the v1 endpoints in 6 months (or re-evaluate).

IDE-4957